Various things: - started on data model for new catalog landing page - created ListingCard which replaces CatalogListing for most uses and contains just the data we need - expanded CatalogListing to contain ItemDetail so that we don't turn around and immediately fetch the item details - nixed a bunch of unused imports - made ItemRating support a mode where it doesn't have the underlying item

+++ src/gwt/client/util/Stars.java
@@ -41,7 +41,7 @@ public abstract class Stars extends FlowPanel
         update();
     }
 
-    protected void updateStarImage (double rating)
+    protected void updateStarImage (float rating)
     {
         int filledStars = (int) (rating * 2);
         String prefix = "/images/ui/stars/" + (_halfSize ? "half" : "full") + "/";
@@ -108,7 +108,7 @@ public abstract class Stars extends FlowPanel
      * This method is called when the mouse pointer enters the widget. The subclasser should
      * update the view here, probably by calling {@link #updateStarImages}.
      */
-    protected abstract void update (double rating);
+    protected abstract void update (float rating);
     
     /**
      * This method is called when the user clicks on the widget.
